reset TAPPING_TERM to 200 for corne
100 is way too low. This manifests itself as modifiers like ESC
(which I hit very often due to vim-like bindings) not registering
as ESC when I accidentally do a long-tap on it. There may be some
permissive hold options fighting me here because it seems quite easy
to trigger when tapping ESC but also happening to hit another key
on either board of the split.