corne: shift, underscore, and colon optimizations
With the day-job switch to Erlang, I find myself frustrated at how
often I need to use my pinkies to use

* Shift
* Underscore

This change swaps the Backspace and Shift bindings for the left
hand and switches right-Shift to be Underscore. This will be
very painful to adjust to but should be more comfortable in
the long run based on how little I like using pinkie shift
and how often I type underscore while programming.

---

Also included is an optimization for Erlang's `=:=` operator which
is used fairly often in guards. Luckily that key in the `_LOWER`
was waiting for a binding - now I don't have to leave `_LOWER` to
type the operator.