Revert "corne: shift, underscore, and colon optimizations"
This reverts commit 1774d9bc2e00117dfcf65a2e4d65b1e1883f772b.

The shift commit I could probably get used to on its own. The
problem is that there's a lot of shuffling between buttons
because of MD_LOWER which is right next to it. Typing symbols
overlaps with the shift modifier in many sequences in Erlang,
function arguments to name one. I think I could pull this off
by switching shift over to the right 1.5u key where KC_SPACE
currently is. MD_RAISE is used much less often (for numbers
and the rarer symbols) so I doubt it would conflict as much.

I'll be procrastinating that change though since it'll be even
harder than this was to get used to.