Remove everything extra in the adjust maps
I accidentally turned on AG_SWAP a while ago and was very confused what
was causing the effect (alt & gui are swapped). I also accidentally
turn on/off audio when I should just be leaving it on.

So I'm getting rid of every mapping in adjust that I don't see the need
for.